CAS-135: Add leaderboard test file and remove unnecessary comments
